Read layered dynamic configuration from filesystem
This pulls in room- and user-specific configuration for a bundle and
layers it on top of the base dynamic configuration for that bundle.

Only reads configuration from the filesystem; this is not hooked up to
Cog's dynamic config API yet (that still only affects the base
configuration), so this will only work for users that have direct
filesystem access to their Relay hosts (i.e., not Hosted Cog).

Fixes operable/cog#850