Drive response error from exit status, not stderr
Previously, a command would be marked as an error if standard error contained any data, regardless of the actual exit code of the command (0 = success, >0 = failure).

This would be especially problematic for commands that wrap existing executables which output additional information on standard error, reserving standard output for the "real" output of the program.

Now, we rely on the exit status only.