Made needsUpdate() devmode aware
Prior to this commit the `-dev` flag was only considered when a command
was run. This meant Relay would consider images up to date at startup
rather than attempting to pull a new image.

This change modifies Relay's staleness checking logic to consider all
images to be stale when `-dev` is in effect during startup, periodic
registry refreshes, and command invocations.

The `operable/circuit-driver` image is omitted from this behavior since
it's integral to Relay's function and doesn't change often.