Exit with a specific exit code when a restart is required
Since f0e0b852 we `Exit()` when the relay is disconnected from Cog as a
trigger to the supervisor that the process has to be restarted.

In order for the supervisor to distinguish between generic errors and
restart requests a new exit code is introduced. This fits well with
Systemd's `RestartForceExitStatus=` directive.