Ping docker for server version
Previously if there was a mismatch between the client and server
versions of the docker library we were using, we didn't have a way to
recover so crashing with an error message was the best we could do. The
user would then have to provide a DOCKER_API_VERSION env var to make
sure the client could connect with the correct version.

In the 1.13.x version of the docker client library a ping call was
introduced, which returns a new header with the server version number.
Now, after creating a client, we can ping the server to check its
version number. If that version number is lower than the client, we can
easily downgrade the client. The only case we still can't recover from
is when the server version is higher than the client, which would
probably require a client lib update in Relay.