Build static Relay executable for Linux
Users have expressed interest in running the bare Relay executable
outside of a Docker environment. Now, we build one on each CI
run (though this isn't yet uploaded to any artifact repository).

In the course of doing this work, it became apparent that the Relay
executable we had been building wasn't actually a statically-linked
executable, and so it would only run on Alpine Linux systems (actually,
a Linux system with musl installed). By using `CGO_ENABLED=0` in our
build step, we can build a static binary, but need to do this on
Ubuntu. (There appear to be additional approaches we could take to build
on Alpine, but it's not really worth it at this point.)

Now, our Docker image build is slightly more involved, in that we build
the executable, and then inject it into a bare-bones container
image. This allows us to not carry around the Go toolchain in the image
layer history. Note that you can still run `make exe` to generate a
binary on your current platform. Running `make docker`, though, will
always generate a Linux executable.

Test images are pushed to the operable/relay-testing repository on
Docker Hub.