Add executable smoke test to Travis
This adapts the basic smoke test we had in Buildkite for Travis
CI. Buildkite has stages, though, and Travis CI doesn't (yet), so we
need to rework things a bit.

Here, we pull the functionality of the old Buildkite scripts
`build_executable.sh` and `smoke_test.sh` into a single shell script
that builds the static relay executable, and then tries to run it on
both Alpine and Ubuntu platforms. Given the constraints in Travis'
build matrix, I've turned the `script` into a multi-step one, and
removed the parallelism. I've also removed the `vet` step, but I think
we can do better by using a separate service (https://ebertapp.io) to
run that instead.